requirements.txt
requests
beautifulsoup4
pyyaml
fake-useragent
feedparser
